Adruino/project_8/requirements.txt
PyQt5==5.15.9
requests==2.31.0
ntplib==0.4.0
pillow==10.0.0
